# api/openapi.yaml
openapi: 3.0.3
info:
  title: glamor daemon API
  version: 0.1.0
  description: |
    HTTP API и WS-события демона glamord (D-02: спека — source of truth,
    Go strict-server + модели через oapi-codegen, TS-типы через
    openapi-typescript — см. уточнение D-02 от 2026-08-16).
    Все мутации идемпотентны (D-12): header `Idempotency-Key`.
    Auth: Bearer-токен (D-08) на всё, кроме /healthz; для /ws — query-параметр.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
        description: динамический порт демона (D-08), см. ~/.glamor/daemon.json
...
paths:
...
  /runs:
    post:
      tags: [runs]
      operationId: createRun
      summary: Создать ран (идемпотентно, D-12)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "200":
          description: повтор Idempotency-Key — возвращён первый ран
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Error"
    get:
      tags: [runs]
      operationId: listRuns
      summary: Список ранов с фильтрами
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
            format: int64
        - name: pipeline_id
          in: query
          schema:
            type: integer
            format: int64
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/RunState"
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: раны
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Error"
...
  /runs/{id}/stop:
    post:
      tags: [runs]
      operationId: stopRun
      summary: Остановить ран (идемпотентно, D-14)
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: остановлен (или уже остановлен)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Error"
  /runs/{id}/resume:
    post:
      tags: [runs]
      operationId: resumeRun
      summary: Ручной рестарт failed/interrupted рана
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: ран возобновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Error"
  /runs/{id}/notes:
    post:
      tags: [runs]
      operationId: createNote
      summary: Queue note к ближайшему событию рана (D-22)
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: заметка поставлена в очередь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        default:
          $ref: "#/components/responses/Error"
...
  /ws:
    get:
      tags: [events]
      operationId: eventsWs
      summary: WebSocket-поток событий журнала (D-04)
      description: |
        Upgrade в WebSocket. Протокол: догон событиями с id > last_event_id,
        сообщение `synced` (граница replay/live), затем live-поток Event'ов.
        Медленный клиент отключается (close 1001, "resync required") и
        переподключается с последним полученным last_event_id.
      parameters:
        - name: run_id
          in: query
          schema:
            type: string
            default: "*"
          description: ран для подписки; "*" — все раны (инбокс гейтов)
        - name: last_event_id
          in: query
          schema:
            type: integer
            format: int64
            default: 0
        - name: token
          in: query
          schema:
            type: string
          description: токен localhost-API (D-08), если настроен
      responses:
        "101":
          description: WebSocket upgrade; далее Event | SyncedMessage
        "401":
          description: неверный/отсутствующий token
components:
  parameters:
...
    RunId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: ключ идемпотентности (D-12); не передан — сервер сгенерирует
  responses:
    Error:
      description: ошибка в едином формате
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            доменный код: not_found, invalid_transition, run_locked,
            gate_already_resolved, duplicate, draining, validation, internal
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          description: |
            контекст ошибки: dirty_checkout → {files: string[]};
            run_locked → {branch: string, run_id: string} (F-02)
...
    RunState:
      type: string
      enum: [draft, running, waiting_gate, succeeded, failed, stopped]
...
    Run:
      type: object
      required: [id, project_id, pipeline_version_id, task_text, base_branch,
                 branch, state, depth, notify_tg, created_at]
      properties:
        id:
          type: string
        project_id:
          type: integer
          format: int64
        pipeline_version_id:
          type: integer
          format: int64
        task_text:
          type: string
        base_branch:
          type: string
        branch:
          type: string
        state:
          $ref: "#/components/schemas/RunState"
        depth:
          type: integer
          format: int64
        notify_tg:
          type: boolean
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
...
    CreateRunRequest:
      type: object
      required: [project_id, pipeline_version_id, task_text]
      properties:
        project_id:
          type: integer
          format: int64
        pipeline_version_id:
          type: integer
          format: int64
        task_text:
          type: string
        base_branch:
          type: string
          description: не указана — default_branch проекта
        branch:
          type: string
          description: не указана — glamor/<slug> от base (D-31)
        depth:
          type: integer
          format: int64
          default: 0
        notify_tg:
          type: boolean
          default: true
        force:
          type: boolean
          default: false
          description: override грязного чекаuta в preflight (D-34, T-10)
...
    Note:
      type: object
      required: [id, run_id, text, consumed, created_at]
      properties:
        id:
          type: integer
          format: int64
        run_id:
          type: string
        stage_id:
          type: integer
          format: int64
          nullable: true
        text:
          type: string
        kind:
          type: string
          enum: [note, steer]
        consumed:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateNoteRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
...